removed base widget
